App: Rearrange screens and add routes
